Add option to show logs from previous pages in the logs dialog
It introduces two new variables to the status:
  logsFilter: function that selects the relevant logs (we use it
    to show only the logs that include the selected datasets).
  logsPageLast: index of the last page that we have retrieved so far.

The Log dialog now contains a new action, a flat button that allows
it to load logs from previous pages.

There are new semantics for the getLogs() function: if it returns
null, then there are no more log pages (so we do not need to request
more). Also, the pages are a positive number, that is, page 1 is the
page previous to the last one (instead of it being page -1).